Parser optimisation & refactor + critical bug fix

+++ Jiten.Parser/SentenceInfo.cs
@@ -3,7 +3,7 @@ namespace Jiten.Parser;
 public class SentenceInfo
 {
     public string Text { get; set; } = string.Empty;
-    public List<(WordInfo word, byte position, byte length)> Words { get; set; } = new();
+    public List<(WordInfo word, int position, int length)> Words { get; set; } = new();
     
     public SentenceInfo(string text)
     {
